Renaming InMemory… classes related to Syncable Store to Default… since they are no longer in memory only implementations
