+++ code/cross-platform-packages/freedom-syncable-store-types/src/internal/types/DefaultAccessControlledFolder.ts
@@ -3,10 +3,10 @@ import type { StaticSyncablePath } from 'freedom-sync-types';
 import type { SyncableStoreBacking } from '../../types/backing/SyncableStoreBacking.ts';
 import type { MutableSyncableStore } from '../../types/MutableSyncableStore.ts';
 import type { SyncTracker } from '../../types/SyncTracker.ts';
-import { InMemoryAccessControlledFolderBase } from './InMemoryAccessControlledFolderBase.ts';
+import { DefaultAccessControlledFolderBase } from './DefaultAccessControlledFolderBase.ts';
 
 // TODO: need to figure out reasonable way of handling partially loaded data, especially for .access-control bundles, since both uploads and downloads are multi-part and async
-export class InMemoryAccessControlledFolder extends InMemoryAccessControlledFolderBase {
+export class DefaultAccessControlledFolder extends DefaultAccessControlledFolderBase {
   constructor({
     store,
     backing,
@@ -22,7 +22,7 @@ export class InMemoryAccessControlledFolder extends InMemoryAccessControlledFold
 
     this.deferredInit_({
       store,
-      makeFolderAccessor: ({ path }) => new InMemoryAccessControlledFolder({ store, backing, path, syncTracker })
+      makeFolderAccessor: ({ path }) => new DefaultAccessControlledFolder({ store, backing, path, syncTracker })
     });
   }
 }
