+++ code/cross-platform-packages/freedom-syncable-store-types/src/internal/types/DefaultFolderStore.ts
@@ -13,140 +13,82 @@ import type { Sha256Hash } from 'freedom-basic-data';
 import { objectEntries } from 'freedom-cast';
 import { ConflictError, generalizeFailureResult, InternalStateError, NotFoundError } from 'freedom-common-errors';
 import { type Trace } from 'freedom-contexts';
-import { generateSha256HashForEmptyString, generateSha256HashFromBuffer, generateSha256HashFromHashesById } from 'freedom-crypto';
+import { generateSha256HashForEmptyString, generateSha256HashFromHashesById } from 'freedom-crypto';
 import { extractPartsFromTimeId, extractPartsFromTrustedTimeId, timeIdInfo, trustedTimeIdInfo } from 'freedom-crypto-data';
 import type {
   DynamicSyncableId,
   StaticSyncablePath,
-  SyncableBundleFileMetadata,
-  SyncableFlatFileMetadata,
+  SyncableFolderMetadata,
   SyncableId,
   SyncableItemType,
   SyncableProvenance
 } from 'freedom-sync-types';
-import { areDynamicSyncableIdsEqual, syncableEncryptedIdInfo, syncableItemTypes } from 'freedom-sync-types';
+import { areDynamicSyncableIdsEqual, invalidProvenance, syncableEncryptedIdInfo } from 'freedom-sync-types';
 import { disableLam } from 'freedom-trace-logging-and-metrics';
 import { flatten } from 'lodash-es';
 import type { SingleOrArray } from 'yaschema';
 
 import type { SyncableStoreBacking } from '../../types/backing/SyncableStoreBacking.ts';
-import type { BundleManagement } from '../../types/BundleManagement.ts';
+import type { FolderManagement } from '../../types/FolderManagement.ts';
 import type { GenerateNewSyncableItemIdFunc } from '../../types/GenerateNewSyncableItemIdFunc.ts';
 import type { LocalItemMetadata } from '../../types/LocalItemMetadata.ts';
-import type { MutableBundleFileAccessor } from '../../types/MutableBundleFileAccessor.ts';
-import type { MutableFileStore } from '../../types/MutableFileStore.ts';
-import type { MutableFlatFileAccessor } from '../../types/MutableFlatFileAccessor.ts';
+import type { MutableAccessControlledFolderAccessor } from '../../types/MutableAccessControlledFolderAccessor.ts';
+import type { MutableFolderStore } from '../../types/MutableFolderStore.ts';
 import type { MutableSyncableItemAccessor } from '../../types/MutableSyncableItemAccessor.ts';
 import type { MutableSyncableStore } from '../../types/MutableSyncableStore.ts';
 import type { SyncableItemAccessor } from '../../types/SyncableItemAccessor.ts';
 import type { SyncTracker } from '../../types/SyncTracker.ts';
-import { generateProvenanceForFileAtPath } from '../../utils/generateProvenanceForFileAtPath.ts';
 import { generateProvenanceForFolderLikeItemAtPath } from '../../utils/generateProvenanceForFolderLikeItemAtPath.ts';
 import { guardIsExpectedType } from '../../utils/guards/guardIsExpectedType.ts';
 import { markSyncableNeedsRecomputeHashAtPath } from '../../utils/markSyncableNeedsRecomputeHashAtPath.ts';
 import { intersectSyncableItemTypes } from '../utils/intersectSyncableItemTypes.ts';
+import { DefaultAccessControlledFolder } from './DefaultAccessControlledFolder.ts';
 import type { FolderOperationsHandler } from './FolderOperationsHandler.ts';
 
-export interface InMemoryBundleBaseConstructorArgs {
-  store: WeakRef<MutableSyncableStore>;
-  backing: SyncableStoreBacking;
-  syncTracker: SyncTracker;
-  folderOperationsHandler: FolderOperationsHandler;
-  path: StaticSyncablePath;
-  supportsDeletion: boolean;
-}
-
-// TODO: rename to DefaultBundleBase in separate PR
-export abstract class InMemoryBundleBase implements MutableFileStore, BundleManagement {
-  public readonly type = 'bundleFile';
+export class DefaultFolderStore implements MutableFolderStore, FolderManagement {
   public readonly path: StaticSyncablePath;
-  public readonly supportsDeletion: boolean;
 
-  protected readonly weakStore_: WeakRef<MutableSyncableStore>;
-  protected readonly syncTracker_: SyncTracker;
-  protected readonly folderOperationsHandler_: FolderOperationsHandler;
+  private readonly syncTracker_: SyncTracker;
 
-  protected readonly backing_: SyncableStoreBacking;
+  private readonly weakStore_: WeakRef<MutableSyncableStore>;
+  private readonly folderOperationsHandler_: FolderOperationsHandler;
+  private readonly makeFolderAccessor_: (args: { path: StaticSyncablePath }) => MutableAccessControlledFolderAccessor;
+
+  private readonly backing_: SyncableStoreBacking;
 
   private needsRecomputeHashCount_ = 0;
 
-  constructor({ store, backing, syncTracker, folderOperationsHandler, path, supportsDeletion }: InMemoryBundleBaseConstructorArgs) {
-    this.supportsDeletion = supportsDeletion;
+  constructor({
+    store,
+    backing,
+    syncTracker,
+    folderOperationsHandler,
+    path,
+    makeFolderAccessor
+  }: {
+    store: WeakRef<MutableSyncableStore>;
+    backing: SyncableStoreBacking;
+    syncTracker: SyncTracker;
+    folderOperationsHandler: FolderOperationsHandler;
+    path: StaticSyncablePath;
+    makeFolderAccessor: (args: { path: StaticSyncablePath }) => MutableAccessControlledFolderAccessor;
+  }) {
     this.weakStore_ = store;
     this.backing_ = backing;
     this.syncTracker_ = syncTracker;
     this.folderOperationsHandler_ = folderOperationsHandler;
     this.path = path;
+    this.makeFolderAccessor_ = makeFolderAccessor;
   }
 
-  // Abstract Methods
+  // MutableFolderStore Methods
 
-  protected abstract computeHash_(trace: Trace, encodedData: Uint8Array): PR<Sha256Hash>;
-  protected abstract decodeData_(trace: Trace, encodedData: Uint8Array): PR<Uint8Array>;
-  protected abstract encodeData_(trace: Trace, rawData: Uint8Array): PR<Uint8Array>;
-  protected abstract makeBundleAccessor_(args: { path: StaticSyncablePath }): MutableBundleFileAccessor;
-  protected abstract makeFlatFileAccessor_(args: { path: StaticSyncablePath }): MutableFlatFileAccessor;
-  protected abstract isEncrypted_(): boolean;
-
-  // MutableFileStore Methods
-
-  public readonly createBinaryFile: MutableFileStore['createBinaryFile'] = makeAsyncResultFunc(
-    [import.meta.filename, 'createBinaryFile'],
-    async (trace: Trace, args): PR<MutableFlatFileAccessor, 'conflict' | 'deleted'> => {
+  public readonly createFolder: MutableFolderStore['createFolder'] = makeAsyncResultFunc(
+    [import.meta.filename, 'createFolder'],
+    async (trace, args): PR<MutableAccessControlledFolderAccessor, 'conflict' | 'deleted'> => {
       switch (args.mode) {
         case 'via-sync':
-          return await this.createPreEncodedBinaryFile_(trace, args.id, args.encodedValue, args.metadata);
-        case undefined:
-        case 'local': {
-          const store = this.weakStore_.deref();
-          if (store === undefined) {
-            return makeFailure(new InternalStateError(trace, { message: 'store was released' }));
-          }
-
-          const encodedData = await this.encodeData_(trace, args.value);
-          /* node:coverage disable */
-          if (!encodedData.ok) {
-            return encodedData;
-          }
-          /* node:coverage enable */
-
-          const getSha256ForItemProvenance = (trace: Trace) => generateSha256HashFromBuffer(trace, encodedData.value);
-
-          const id = await this.folderOperationsHandler_.generateNewSyncableItemId(trace, {
-            id: args.id,
-            parentPath: this.path,
-            getSha256ForItemProvenance
-          });
-          /* node:coverage disable */
-          if (!id.ok) {
-            return id;
-          }
-          /* node:coverage enable */
-
-          const newPath = this.path.append(id.value);
-
-          const provenance = await generateProvenanceForFileAtPath(trace, store, { path: newPath, getSha256ForItemProvenance });
-          if (!provenance.ok) {
-            return provenance;
-          }
-
-          return await this.createPreEncodedBinaryFile_(trace, id.value, encodedData.value, {
-            type: 'flatFile',
-            provenance: provenance.value,
-            // Flat file encryption always matches the parent bundle or folder
-            encrypted: this.isEncrypted_()
-          });
-        }
-      }
-    }
-  );
-
-  public readonly createBundleFile: MutableFileStore['createBundleFile'] = makeAsyncResultFunc(
-    [import.meta.filename, 'createBundleFile'],
-    async (trace, args): PR<MutableBundleFileAccessor, 'conflict' | 'deleted'> => {
-      switch (args.mode) {
-        case 'via-sync':
-          return await this.createPreEncodedBundleFile_(trace, args.id, args.metadata);
+          return await this.createPreEncodedFolder_(trace, args.id, args.metadata);
         case undefined:
         case 'local': {
           const store = this.weakStore_.deref();
@@ -170,11 +112,23 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
             return provenance;
           }
 
-          return await this.createPreEncodedBundleFile_(trace, id.value, {
-            type: 'bundleFile',
+          const folder = await this.createPreEncodedFolder_(trace, id.value, {
+            type: 'folder',
             provenance: provenance.value,
-            encrypted: this.isEncrypted_()
+            encrypted: true
           });
+          if (!folder.ok) {
+            return folder;
+          }
+
+          // Initializing folder
+
+          const accessInitialized = await folder.value.initialize(trace);
+          if (!accessInitialized.ok) {
+            return accessInitialized;
+          }
+
+          return makeSuccess(folder.value);
         }
       }
     }
@@ -192,10 +146,6 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
         id = staticId.value;
       }
 
-      if (!this.supportsDeletion) {
-        return makeFailure(new InternalStateError(trace, { message: `Deletion is not supported in ${this.path.toString()}` }));
-      }
-
       const removePath = this.path.append(id);
 
       // Checking that the requested file exists in the backing
@@ -206,7 +156,7 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
       } else if (!exists.value) {
         return makeFailure(
           new NotFoundError(trace, {
-            message: `No file found for ID: ${id} in ${this.path.toString()}`,
+            message: `No folder found for ID: ${id} in ${this.path.toString()}`,
             errorCode: 'not-found'
           })
         );
@@ -234,6 +184,9 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
         return markedNeedsRecomputedHash;
       }
       /* node:coverage enable */
+
+      DEV: debugTopic('SYNC', (log) => log(`Notifying folderRemoved for folder ${removePath.toString()}`));
+      this.syncTracker_.notify('folderRemoved', { path: removePath });
 
       return makeSuccess(undefined);
     }
@@ -300,13 +253,7 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
 
       const guards = await allResults(trace, [
         this.guardNotDeleted_(trace, getPath, 'deleted'),
-        guardIsExpectedType(
-          trace,
-          getPath,
-          metadata.value,
-          intersectSyncableItemTypes(expectedType, syncableItemTypes.exclude('folder')),
-          'wrong-type'
-        )
+        guardIsExpectedType(trace, getPath, metadata.value, intersectSyncableItemTypes(expectedType, 'folder'), 'wrong-type')
       ]);
       if (!guards.ok) {
         return guards;
@@ -322,7 +269,7 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
   public readonly staticToDynamicId = (trace: Trace, id: SyncableId): PR<DynamicSyncableId> =>
     this.folderOperationsHandler_.staticToDynamicId(trace, id);
 
-  // FileStore Methods
+  // FolderStore Methods
 
   public readonly getHash = makeAsyncResultFunc(
     [import.meta.filename, 'getHash'],
@@ -355,9 +302,9 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
         /* node:coverage enable */
 
         if (this.needsRecomputeHashCount_ === needsRecomputeHashCount) {
-          const updatedHash = await this.backing_.updateLocalMetadataAtPath(trace, this.path, { hash: hash.value });
-          if (!updatedHash.ok) {
-            return generalizeFailureResult(trace, updatedHash, ['not-found', 'wrong-type']);
+          const updatedMetadata = await this.backing_.updateLocalMetadataAtPath(trace, this.path, { hash: hash.value });
+          if (!updatedMetadata.ok) {
+            return generalizeFailureResult(trace, updatedMetadata, ['not-found', 'wrong-type']);
           }
 
           return makeSuccess(hash.value);
@@ -370,7 +317,7 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
     [import.meta.filename, 'getHashesById'],
     async (trace: Trace, { recompute = false }: { recompute?: boolean } = {}): PR<Partial<Record<SyncableId, Sha256Hash>>> => {
       // Deleted files are already filtered out using getIds
-      const ids = await this.getIds(trace);
+      const ids = await this.getIds(trace, { type: 'folder' });
       /* node:coverage disable */
       if (!ids.ok) {
         return ids;
@@ -477,28 +424,13 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
     [import.meta.filename, 'getIds'],
     async (trace: Trace, options?: { type?: SingleOrArray<SyncableItemType> }): PR<SyncableId[]> => {
       const ids = await this.backing_.getIdsInPath(trace, this.path, {
-        type: intersectSyncableItemTypes(options?.type, syncableItemTypes.exclude('folder'))
+        type: intersectSyncableItemTypes(options?.type, 'folder')
       });
       if (!ids.ok) {
         return generalizeFailureResult(trace, ids, ['not-found', 'wrong-type']);
       }
 
-      const nonDeletedIds = await this.filterOutDeletedIds_(trace, ids.value);
-      if (!nonDeletedIds.ok) {
-        return nonDeletedIds;
-      }
-
-      const metadataById = await this.backing_.getMetadataByIdInPath(trace, this.path, new Set(nonDeletedIds.value));
-      if (!metadataById.ok) {
-        return generalizeFailureResult(trace, metadataById, ['not-found', 'wrong-type']);
-      }
-
-      const isEncrypted = this.isEncrypted_();
-      const idsWithMatchingEncryptionMode = objectEntries(metadataById.value)
-        .filter(([_id, metadata]) => metadata?.encrypted === isEncrypted)
-        .map(([id, _metadata]) => id);
-
-      return makeSuccess(idsWithMatchingEncryptionMode);
+      return await this.filterOutDeletedIds_(trace, ids.value);
     }
   );
 
@@ -511,23 +443,17 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
     ): PR<SyncableItemAccessor & { type: T }, 'deleted' | 'not-found' | 'wrong-type'> => await this.getMutable(trace, id, expectedType)
   );
 
-  public readonly getProvenance = makeAsyncResultFunc([import.meta.filename, 'getProvenance'], async (trace): PR<SyncableProvenance> => {
-    const metadata = await this.backing_.getMetadataAtPath(trace, this.path);
-    if (!metadata.ok) {
-      return generalizeFailureResult(trace, metadata, ['not-found', 'wrong-type']);
-    }
-
-    const provenance = metadata.value.provenance;
-
-    return makeSuccess(provenance);
-  });
+  public readonly getProvenance = makeAsyncResultFunc(
+    [import.meta.filename, 'getProvenance'],
+    async (_trace): PR<SyncableProvenance> => makeSuccess(invalidProvenance)
+  );
 
   public readonly markNeedsRecomputeHash = makeAsyncResultFunc(
     [import.meta.filename, 'markNeedsRecomputeHash'],
     async (trace): PR<undefined> => {
-      const updatedHash = await this.backing_.updateLocalMetadataAtPath(trace, this.path, { hash: undefined });
-      if (!updatedHash.ok) {
-        return generalizeFailureResult(trace, updatedHash, ['not-found', 'wrong-type']);
+      const updatedMetadata = await this.backing_.updateLocalMetadataAtPath(trace, this.path, { hash: undefined });
+      if (!updatedMetadata.ok) {
+        return generalizeFailureResult(trace, updatedMetadata, ['not-found', 'wrong-type']);
       }
 
       this.needsRecomputeHashCount_ += 1;
@@ -548,14 +474,14 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
   );
 
   public readonly ls = makeAsyncResultFunc([import.meta.filename, 'ls'], async (trace): PR<string[]> => {
-    const hashesByFileId = await this.getHashesById(trace);
-    if (!hashesByFileId.ok) {
-      return hashesByFileId;
+    const hashesByFolderId = await this.getHashesById(trace);
+    if (!hashesByFolderId.ok) {
+      return hashesByFolderId;
     }
 
     const recursiveLs = await allResultsMapped(
       trace,
-      objectEntries(hashesByFileId.value).sort((a, b) => a[0].localeCompare(b[0])),
+      objectEntries(hashesByFolderId.value).sort((a, b) => a[0].localeCompare(b[0])),
       {},
       async (trace, [itemId, hash]): PR<string[]> => {
         const getPath = this.path.append(itemId);
@@ -568,9 +494,7 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
 
         const output: string[] = [`${dynamicId.ok ? JSON.stringify(dynamicId.value) : itemId}: ${hash}`];
         switch (metadata.value.type) {
-          case 'folder':
-            break; // This won't happen
-          case 'bundleFile': {
+          case 'folder': {
             const itemAccessor = this.makeItemAccessor_(getPath, metadata.value.type);
             const fileLs = await itemAccessor.ls(trace);
             if (!fileLs.ok) {
@@ -581,8 +505,9 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
 
             break;
           }
+          case 'bundleFile':
           case 'flatFile':
-            break; // Nothing to do
+            break; // These won't happen
         }
 
         return makeSuccess(output);
@@ -595,14 +520,10 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
     return makeSuccess(flatten(recursiveLs.value));
   });
 
-  // BundleManagement Methods
+  // FileStoreManagementAccessor Methods
 
-  public readonly sweep = makeAsyncResultFunc([import.meta.filename, 'sweep'], async (trace: Trace): PR<undefined> => {
-    if (!this.supportsDeletion) {
-      return makeSuccess(undefined);
-    }
-
-    const allItemIds = await this.backing_.getIdsInPath(trace, this.path, { type: syncableItemTypes.exclude('folder') });
+  public async sweep(trace: Trace): PR<undefined> {
+    const allItemIds = await this.backing_.getIdsInPath(trace, this.path, { type: 'folder' });
     if (!allItemIds.ok) {
       return generalizeFailureResult(trace, allItemIds, ['not-found', 'wrong-type']);
     } else if (allItemIds.value.length === 0) {
@@ -627,12 +548,12 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
       return generalizeFailureResult(trace, deletedInBacking, ['not-found', 'wrong-type']);
     }
 
-    const subBundleIds = await this.backing_.getIdsInPath(trace, this.path, { type: 'bundleFile' });
-    if (!subBundleIds.ok) {
-      return generalizeFailureResult(trace, subBundleIds, ['not-found', 'wrong-type']);
+    const subFolderIds = await this.backing_.getIdsInPath(trace, this.path, { type: 'folder' });
+    if (!subFolderIds.ok) {
+      return generalizeFailureResult(trace, subFolderIds, ['not-found', 'wrong-type']);
     }
-    const recursivelySwept = await allResultsMapped(trace, subBundleIds.value, {}, async (trace, bundleId) => {
-      const itemAccessor = this.makeMutableItemAccessor_(this.path.append(bundleId), 'bundleFile');
+    const recursivelySwept = await allResultsMapped(trace, subFolderIds.value, {}, async (trace, folderId) => {
+      const itemAccessor = this.makeMutableItemAccessor_(this.path.append(folderId), 'folder');
       return await itemAccessor.sweep(trace);
     });
     /* node:coverage disable */
@@ -642,77 +563,17 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
     /* node:coverage enable */
 
     return makeSuccess(undefined);
-  });
+  }
 
   // Private Methods
 
-  private readonly createPreEncodedBinaryFile_ = makeAsyncResultFunc(
-    [import.meta.filename, 'createPreEncodedBinaryFile_'],
-    async (
-      trace: Trace,
-      id: SyncableId,
-      encodedData: Uint8Array,
-      metadata: SyncableFlatFileMetadata & LocalItemMetadata
-    ): PR<MutableFlatFileAccessor, 'conflict' | 'deleted'> => {
-      const newPath = this.path.append(id);
-
-      const exists = await this.backing_.existsAtPath(trace, newPath);
-      if (!exists.ok) {
-        return exists;
-      } else if (exists.value) {
-        return makeFailure(new ConflictError(trace, { message: `${newPath.toString()} already exists`, errorCode: 'conflict' }));
-      }
-
-      const guards = await this.guardNotDeleted_(trace, newPath, 'deleted');
-      if (!guards.ok) {
-        return guards;
-      }
-
-      const hash = await this.computeHash_(trace, encodedData);
-      /* node:coverage disable */
-      if (!hash.ok) {
-        return hash;
-      }
-      /* node:coverage enable */
-
-      // Never using a hash from metadata
-      metadata.hash = hash.value;
-
-      const createdFlatFile = await this.backing_.createBinaryFileWithPath(trace, newPath, {
-        data: encodedData,
-        metadata
-      });
-      if (!createdFlatFile.ok) {
-        return generalizeFailureResult(trace, createdFlatFile, ['not-found', 'wrong-type']);
-      }
-
-      const itemAccessor = this.makeItemAccessor_(newPath, 'flatFile');
-
-      const marked = await itemAccessor.markNeedsRecomputeHash(trace);
-      /* node:coverage disable */
-      if (!marked.ok) {
-        return marked;
-      }
-      /* node:coverage enable */
-
-      DEV: debugTopic('SYNC', (log) => log(`Notifying needsSync for flat file ${newPath.toString()}`));
-      this.syncTracker_.notify('needsSync', {
-        type: 'flatFile',
-        path: newPath,
-        hash: hash.value
-      });
-
-      return makeSuccess(itemAccessor);
-    }
-  );
-
-  public readonly createPreEncodedBundleFile_ = makeAsyncResultFunc(
-    [import.meta.filename, 'createPreEncodedBundleFile_'],
+  private readonly createPreEncodedFolder_ = makeAsyncResultFunc(
+    [import.meta.filename, 'createPreEncodedFolder_'],
     async (
       trace,
       id: SyncableId,
-      metadata: SyncableBundleFileMetadata & LocalItemMetadata
-    ): PR<MutableBundleFileAccessor, 'conflict' | 'deleted'> => {
+      metadata: SyncableFolderMetadata & LocalItemMetadata
+    ): PR<DefaultAccessControlledFolder, 'conflict' | 'deleted'> => {
       const newPath = this.path.append(id);
 
       const exists = await this.backing_.existsAtPath(trace, newPath);
@@ -727,10 +588,17 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
         return guards;
       }
 
-      const store = this.weakStore_.deref();
-      if (store === undefined) {
-        return makeFailure(new InternalStateError(trace, { message: 'store was released' }));
+      const createdFolder = await this.backing_.createFolderWithPath(trace, newPath, { metadata });
+      if (!createdFolder.ok) {
+        return generalizeFailureResult(trace, createdFolder, ['not-found', 'wrong-type']);
       }
+
+      const folder = new DefaultAccessControlledFolder({
+        store: this.weakStore_,
+        backing: this.backing_,
+        syncTracker: this.syncTracker_,
+        path: newPath
+      });
 
       const hash = await generateSha256HashFromHashesById(trace, {});
       /* node:coverage disable */
@@ -739,41 +607,30 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
       }
       /* node:coverage enable */
 
-      // Never using a hash from metadata
-      metadata.hash = hash.value;
-
-      const createdBundleFile = await this.backing_.createFolderWithPath(trace, newPath, { metadata });
-      if (!createdBundleFile.ok) {
-        return generalizeFailureResult(trace, createdBundleFile, ['not-found', 'wrong-type']);
-      }
-
-      const itemAccessor = this.makeMutableItemAccessor_(newPath, 'bundleFile');
-
-      const marked = await itemAccessor.markNeedsRecomputeHash(trace);
+      const marked = await folder.markNeedsRecomputeHash(trace);
       /* node:coverage disable */
       if (!marked.ok) {
         return marked;
       }
       /* node:coverage enable */
 
-      DEV: debugTopic('SYNC', (log) => log(`Notifying needsSync for bundle ${newPath.toString()}`));
+      DEV: debugTopic('SYNC', (log) => log(`Notifying folderAdded for folder ${newPath.toString()}`));
+      this.syncTracker_.notify('folderAdded', { path: newPath });
+
+      DEV: debugTopic('SYNC', (log) => log(`Notifying needsSync for folder ${newPath.toString()}`));
       this.syncTracker_.notify('needsSync', {
-        type: 'bundleFile',
+        type: 'folder',
         path: newPath,
         hash: hash.value
       });
 
-      return makeSuccess(itemAccessor);
+      return makeSuccess(folder);
     }
   );
 
   private readonly filterOutDeletedIds_ = makeAsyncResultFunc(
     [import.meta.filename, 'filterOutDeletedIds_'],
     async (trace, encodedIds: SyncableId[]) => {
-      if (!this.supportsDeletion) {
-        return makeSuccess(encodedIds);
-      }
-
       return await allResultsReduced(
         trace,
         encodedIds,
@@ -799,10 +656,6 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
   private readonly guardNotDeleted_ = makeAsyncResultFunc(
     [import.meta.filename, 'guardNotDeleted_'],
     async <ErrorCodeT extends string>(trace: Trace, path: StaticSyncablePath, errorCode: ErrorCodeT): PR<undefined, ErrorCodeT> => {
-      if (!this.supportsDeletion) {
-        return makeSuccess(undefined);
-      }
-
       const isDeleted = await this.folderOperationsHandler_.isPathMarkedAsDeleted(trace, path);
       if (!isDeleted.ok) {
         return isDeleted;
@@ -824,13 +677,13 @@ export abstract class InMemoryBundleBase implements MutableFileStore, BundleMana
   ): MutableSyncableItemAccessor & { type: T } {
     switch (itemType) {
       case 'folder':
-        throw new Error("Folders can't be managed by InMemoryBundleBase");
+        return this.makeFolderAccessor_({ path }) as any as MutableSyncableItemAccessor & { type: T };
 
       case 'bundleFile':
-        return this.makeBundleAccessor_({ path }) as any as MutableSyncableItemAccessor & { type: T };
+        throw new Error("Bundles can't be managed by DefaultFolderStore");
 
       case 'flatFile':
-        return this.makeFlatFileAccessor_({ path }) as any as MutableSyncableItemAccessor & { type: T };
+        throw new Error("Flat files can't be managed by DefaultFolderStore");
     }
   }
 }
