+++ code/cross-platform-packages/freedom-syncable-store-types/src/internal/types/DefaultMutableFileAccessorBase.ts
@@ -2,7 +2,7 @@ import type { StaticSyncablePath } from 'freedom-sync-types';
 
 import type { MutableSyncableStore } from '../../types/MutableSyncableStore.ts';
 
-export abstract class InMemoryMutableFileAccessorBase {
+export abstract class DefaultMutableFileAccessorBase {
   public abstract type: 'bundleFile' | 'flatFile';
   public readonly path: StaticSyncablePath;
