+++ code/cross-platform-packages/freedom-syncable-store-types/src/internal/types/__tests__/hashes.test.ts
@@ -10,8 +10,8 @@ import { encId, storageRootIdInfo } from 'freedom-sync-types';
 import { expectOk } from 'freedom-testing-tools';
 
 import { makeCryptoServiceForTesting } from '../../../__test_dependency__/makeCryptoServiceForTesting.ts';
+import { DefaultSyncableStore } from '../../../types/DefaultSyncableStore.ts';
 import { InMemorySyncableStoreBacking } from '../../../types/in-memory-backing/InMemorySyncableStoreBacking.ts';
-import { InMemorySyncableStore } from '../../../types/InMemorySyncableStore.ts';
 import { createBinaryFileAtPath } from '../../../utils/create/createBinaryFileAtPath.ts';
 import { createBundleFileAtPath } from '../../../utils/create/createBundleFileAtPath.ts';
 import { createFolderAtPath } from '../../../utils/create/createFolderAtPath.ts';
@@ -23,7 +23,7 @@ describe('hashes', () => {
   let cryptoKeys!: PrivateCombinationCryptoKeySet;
   let cryptoService!: CryptoService;
   let storeBacking!: InMemorySyncableStoreBacking;
-  let store!: InMemorySyncableStore;
+  let store!: DefaultSyncableStore;
 
   const storageRootId = storageRootIdInfo.make('test');
 
@@ -40,7 +40,7 @@ describe('hashes', () => {
     expectOk(provenance);
 
     storeBacking = new InMemorySyncableStoreBacking({ provenance: provenance.value });
-    store = new InMemorySyncableStore({ storageRootId, backing: storeBacking, cryptoService, provenance: provenance.value });
+    store = new DefaultSyncableStore({ storageRootId, backing: storeBacking, cryptoService, provenance: provenance.value });
 
     expectOk(await initializeRoot(trace, store));
   });
