+++ code/cross-platform-packages/freedom-syncable-store-types/src/types/exports.ts
@@ -2,6 +2,7 @@ export * from './AccessControlledFolderAccessor.ts';
 export * from './backing/exports.ts';
 export * from './BundleFileAccessor.ts';
 export * from './BundleManagement.ts';
+export * from './DefaultSyncableStore.ts';
 export * from './FileAccessor.ts';
 export * from './FileAccessorBase.ts';
 export * from './FileStore.ts';
@@ -10,7 +11,6 @@ export * from './FolderManagement.ts';
 export * from './FolderStore.ts';
 export * from './GenerateNewSyncableItemIdFunc.ts';
 export * from './in-memory-backing/exports.ts';
-export * from './InMemorySyncableStore.ts';
 export * from './MutableAccessControlledFolderAccessor.ts';
 export * from './MutableBundleFileAccessor.ts';
 export * from './MutableFileAccessor.ts';
